Kernel/aarch64: Add packed attribute to RPi Mailbox messages
KUBSAN was causing a crash because these structs were not correctly
aligned. Instead of manually adding the aligned attribute, this solves
the problem by declaring each of the message structs as packed, after
which the compiler is able to figure out the correct layout/alignment.
Note: We do need the aligned attribute on the variable on the stack, as
the message must be 16 byte aligned.